+++ R/RcppExports.R
@@ -85,6 +85,10 @@ node_set_attr <- function(node_, name, value, nsMap) {
     invisible(.Call(`_xml2_node_set_attr`, node_, name, value, nsMap))
 }
 
+node_remove_attr <- function(node_, name, nsMap) {
+    invisible(.Call(`_xml2_node_remove_attr`, node_, name, nsMap))
+}
+
 node_children <- function(node, onlyNode = TRUE) {
     .Call(`_xml2_node_children`, node, onlyNode)
 }