+++ R/xml_attr.R
@@ -118,7 +118,7 @@ xml_attrs.xml_nodeset <- function(x, ns = character()) {
 #' @export
 `xml_attr<-.xml_node` <- function(x, attr, ns = character(), value) {
   if (is.null(value)) {
-    node_set_attr(x$node, name = attr, nsMap = ns, "")
+    node_remove_attr(x$node, name = attr, nsMap = ns)
   } else {
     value <- as.character(value)
     node_set_attr(x$node, name = attr, nsMap = ns, value)